+++ Assets/BlessYou/Scripts/Player/Jikken.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e412eba479a6c84418e1fd4f661f5119
+guid: 07a7557f21ff14fa8a3b7bf454099cd8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
