+++ Assets/BlessYou/Scripts/Player/Player.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 297b8fde85a8a9142bf5b8ef9489ef3a
+guid: 6bcae5c32927c40c9b5d60c9b42bba56
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
